chat/requirements.txt
flask==2.2.2
requests==2.28.1
urllib3==1.26.12
flask-mysql==1.5.2
psycopg2-binary
